==> eunok_hwang/03_naver_api_geocoding.ipynb <==
# **03. 네이버 Geocoding API 사용 위·경도 구하기**

**사용 후 Client ID와 Client Secret값을 삭제했으므로 아래 코드는 작동하지 않습니다**

import requests
import pandas as pd

hospital = pd.read_excel("전국병의원_202306.xlsx")
hospital = hospital.loc[hospital['시도코드명'].isin(['서울', '경기', '인천']), ['요양기관명', '종별코드명', '시도코드명', '주소']]

hospital['종별코드명'] = hospital['종별코드명'].replace(['상급종합'], '3차')
hospital['종별코드명'] = hospital['종별코드명'].replace(['종합병원', '병원', '한방병원', '치과병원'], '2차')
hospital['종별코드명'] = hospital['종별코드명'].replace(['의원', '치과의원', '한의원', '보건소', '보건지소', '보건진료소', '보건의료원'], '1차')

hospital = hospital.loc[hospital['종별코드명'].isin(['1차', '2차', '3차'])]

client_id = "(개인정보)"
client_secret = "(개인정보)"

addresses = hospital['주소'].apply(lambda x: x.split(",")[0]).to_list()

location_list = []
for address in addresses: 
    endpoint = "https://naveropenapi.apigw.ntruss.com/map-geocode/v2/geocode"
    url = f"{endpoint}?query={address}"

    headers = {
        "X-NCP-APIGW-API-KEY-ID": client_id,
        "X-NCP-APIGW-API-KEY": client_secret,
        }

    res = requests.get(url, headers=headers)
    address_data = res.json().get('addresses')  
    # .get(): 딕셔너리에서 주어진 키에 해당하는 값을 가져오고 만약 키가 존재하지 않는 경우에도 오류 없이 None을 반환한다

    if address_data:
        lng = address_data[0]['x']
        lat = address_data[0]['y']
        location = (float(lat), float(lng))
    else:
        location = (0.0, 0.0)
        
    location_list.append(location)

hospital['location'] = location_list

hospital[hospital['location'] == (0,0)]

hospital.to_csv("hospital.csv", index=False)

subway = pd.read_csv("subway.csv")

client_id = "(개인정보)"
client_secret = "(개인정보)"

addresses = subway['역사도로명주소'].to_list()

location_list = []
for address in addresses: 
    endpoint = "https://naveropenapi.apigw.ntruss.com/map-geocode/v2/geocode"
    url = f"{endpoint}?query={address}"

    headers = {
        "X-NCP-APIGW-API-KEY-ID": client_id,
        "X-NCP-APIGW-API-KEY": client_secret,
        }

    res = requests.get(url, headers=headers)
    address_data = res.json().get('addresses')  

    if address_data:
        lng = address_data[0]['x']
        lat = address_data[0]['y']
        location = (float(lat), float(lng))
    else:
        location = (0.0, 0.0)
        
    location_list.append(location)

subway['fixed_location'] = location_list

errored = subway['fixed_location'] == (0.0, 0.0)
subway.loc[errored, 'fixed_location'] = subway.loc[errored, 'location']

subway.to_csv("fixed_subway.csv", encoding='UTF-8', index=False)
